refactor(axum): extract validate_port into shared handlers/port_utils
The port SSRF guard was a private function in chat_api.rs. Moving it to
handlers/port_utils.rs makes it reusable by the new agent SSE handler
without duplicating the range check and running-server validation logic.

- New: crates/gglib-axum/src/handlers/port_utils.rs (pub(crate))
- chat_api.rs now imports validate_port from handlers::port_utils
- handlers/mod.rs registers the new submodule

Part of #266 / epic #247
