+++ crates/gglib-axum/src/handlers/mod.rs
@@ -14,5 +14,6 @@ pub mod servers;
 pub mod settings;
 pub mod setup;
 pub mod verification;
+pub mod port_utils;
 pub mod voice;
 pub mod voice_ws;
